Extending power calculations to support noninferiority testing
Beyond simple comparative tests, there exist more flexible hypotheses such as:
    H0: epsilon < delta vs H1 epsilon > delta
with epsilon the true difference in means and delta any real number.
One very useful setup for online experimentation is called noninferiority
testing. When introducing a new feature that is not meant to improve business
level metrics, teams might still want to guard against any significant harm to
these metrics. One way to do so is to setup up the following hypothesis:
    H0: epsilon < -delta vs H1: epsilon > -delta
where delta is a positive real number called "acceptable cost".

Since we don't want to expose the full intricacies of how to define the null hypothesis
for nonsignificance test we augmented the UI for the power calculator with
a "Use noninferiority" tickbox which then prompts the user for the acceptable
cost expressed either as a relative change compared to the control rate or as an
absolute impact per day (the change times the number of visitors per day).
We chose these two ways of expressing the change as they are more natural to people
compared to asking for the absolute difference in means. Therefore we needed to
extend the power calculation utilities to handle this new use cases.

Internally we map back the acceptable cost into an absolute change. That means that
there is some added variance that we discard. For typical uses cases (relative changes
in the order of a couple of percents) that is fine as shown by the tests suite. However,
we haven't tested the impact of the added variance on larger relative changes.