Modifying the power utilities to support more flexible hypothesis tests (#10)
##Modifying the power utilities to support more flexible hypothesis tests

This introduces the possibility to specify more flexible hypothesis tests
for each of the power utilities. The calculations used to default to the following:
H0: epsilon = 0 vs H1: epsilon != 0, with epsilon the difference in means between
treatment and control. Now, via the introduction of the `alternative` and `mu`
parameters, you can speficiy hypothesis tests such as:
 - (alternative='greater', mu=delta) => H0: epsilon <= delta vs epsilon > delta
 - (alternative='lower', mu=defta)   => H0: epsilon >= delta vs epsilon < delta
 - (mu=delta)                        => H0: epsilon = delta vs epsilon != delta

##Several simplifications were introduced for consistency:
 - Across all utilities I defaulted to using a normal approximation for
the test statistics (instead of the exact t-distribution). Indeed, `solveforsample_{T,G}test
was already making this assumption (by fixing the degrees of freedom to a large number)
so the simplification should make the utilities more consistent
 - In `solveforpower_Gtest` I opted for a test with unequal variances and didn't
fallback to the equal variance assumption when mu=0 (for simplicity)

There was also a mistake in `solveforeffectsize_Ttest` where we would return the square
of the correct value.

##A test suite was introduced:
This is far from being exhaustive but should get us started.
Specifically I test each utility by using their output as parameters for simulated experiments. I then compare the observed power obtained with the theoretical power doing an equivalence test with a 0.01 margin (a two one-sided test). The margin is chosen so that the equivalence test has adequate power (0.999) and false positive rate (0.001) without being computationally prohibitive (30000 experiments, as given by -in R with the TOSTER library-:
ceiling(powerTOSTone.raw(alpha = 0.001, statistical_power = 0.999, sd = sqrt(0.8*0.2),
    low_eqbound = -0.015, high_eqbound = 0.015)))

I test most of the utilities by simulating bernoulli distributed metrics (except
the ones which are radically different for binary vs continuous metrics) instead
of normally distributed ones for speed (one draw of binomial accounts for a whole
experiment)


##Beyond simple comparative tests, there exist more flexible hypotheses such as:
    H0: epsilon < delta vs H1 epsilon > delta
with epsilon the true difference in means and delta any real number.
One very useful setup for online experimentation is called noninferiority
testing. When introducing a new feature that is not meant to improve business
level metrics, teams might still want to guard against any significant harm to
these metrics. One way to do so is to setup up the following hypothesis:
    H0: epsilon < -delta vs H1: epsilon > -delta
where delta is a positive real number called "acceptable cost".

Since we don't want to expose the full intricacies of how to define the null hypothesis
for nonsignificance test we augmented the UI for the power calculator with
a "Use noninferiority" tickbox which then prompts the user for the acceptable
cost expressed either as a relative change compared to the control rate or as an
absolute impact per day (the change times the number of visitors per day).
We chose these two ways of expressing the change as they are more natural to people
compared to asking for the absolute difference in means. Therefore we needed to
extend the power calculation utilities to handle this new use cases.

Internally we map back the acceptable cost into an absolute change. That means that
there is some added variance that we discard. For typical uses cases (relative changes
in the order of a couple of percents) that is fine as shown by the tests suite. However,
we haven't tested the impact of the added variance on larger relative changes.

##A couple of fixes/improvements:
 - change the math method so that the value of visitors_per_day passed
from the interface is considered as the total number of visitors per day
across variants
 - validate the inputs so that the "acceptable cost" is not smaller than
the true effect size (it's impossible to not harm by a number smaller
than the true effect size)
 - have the view return `-threshold` instead of `threshold`: it used to
be done in the math module but it makes more sense to have the math module
be dumb about its inputs and have the view do the translation between the
user friendly input and what the math module expects when doing noninferiority
testing