We are replacing impact block with non inferiority when that's the case. - Changing base change noninferiority impact only when it is absolute per day - New tests for non inferiority new field (Expected Change) and bugfixes
Simplify selection of the effect size under the alternative

For non-inferiority testing, not only do people have to select the acceptable
cost, but they also have to specify what is the effect size under the alternative.
However, we can simplify everyone's life by breaking down what are the three
plausible scenarios under the alternative, given we already have the acceptable cost:
  - if we don't expect any change under the alternative then the effect size is
set to 0
  - if we expect a degradation then the effect size under the alternative is set
to a decrease of half of the acceptable cost
  - if we expect an improvement (but for some reason still want to perform a
non-inferiority test) then the effect size under the alternative is set to
the acceptatble cost (we hypothesize that people wouldn't care for an improvement
smaller than the acceptable cost and would in this case rather select the "no change"
case.

We also moved the acceptable cost selection to the impact box (rightmost column)
to improve the UX (no backing data on that though)