feat: changes the display of the base rate to be consistent between modes
During the last refactoring of the power calculator, one of the rules we
imposed ourselves was to respect user input. This is, if the user wrote "400",
after switching it would still be "400", but changing the equivalent internally
from "400 units" to "400%".

The best solution is to remove the % symbol in the binomial metrics. That way,
we don't need to transform the internal values for the user input to stay the
same, and will avoid user confusion.